relocate logic for general-purpose HDF5 outputs
