+++ src/io/io.cpp
@@ -16,7 +16,6 @@
 #endif  // HDF5
 #include "../grid/grid3D.h"
 #include "../io/FieldWriter.h"
-#include "../io/WriterManager.h"
 #include "../io/io.h"
 #include "../utils/cuda_utilities.h"
 #include "../utils/hydro_utilities.h"
@@ -141,180 +140,6 @@ void Write_Data(Grid3D &G, struct Parameters P, int nfile, const io::WriterManag
 #ifdef MPI_CHOLLA
   MPI_Barrier(world);
 #endif
-}
-
-/*! does the heavy-lifting of writing fields to hdf5 files.
- *
- *  \todo
- *  The initial implementation goes out of its way to retain all of the historical
- *  distinctions that occur between regular outputs and float32 outputs. We should
- *  really cut down on these differences!
- */
-template <bool ForceF32Output>
-void Write_Fields_to_HDF5_helper_(const std::string &filename, Grid3D &G, const io::DatasetSpec &dataset_spec,
-                                  io::LazyScratchBuf &lazy_scratch_buf)
-{
-#ifdef HDF5
-  // get the value-type of the dataset buffers
-  using T = std::conditional_t<ForceF32Output, float, Real>;
-
-  const Header &H = G.H;
-  bool is_3D      = H.nx > 1 and H.ny > 1 and H.nz > 1;
-
-  if (ForceF32Output and not is_3D) {
-    // this approximates historical data... Historically we would actually make a file
-    // but not record fields to it, but this is close enough!
-    return;
-  }
-
-  // Create a new file using default properties.
-  hid_t file_id = H5Fcreate(filename.data(), H5F_ACC_TRUNC, H5P_DEFAULT, H5P_DEFAULT);
-
-  // Write the header (file attributes)
-  G.Write_Header_HDF5(file_id);
-
-  // Allocate necessary buffers
-  int nx_dset = H.nx_real;
-  int ny_dset = H.ny_real;
-  int nz_dset = H.nz_real;
-  #ifdef MHD
-  size_t buffer_size = (nx_dset + 1) * (ny_dset + 1) * (nz_dset + 1);
-  #else
-  size_t buffer_size = nx_dset * ny_dset * nz_dset;
-  #endif
-  T *dev_dataset_buf  = lazy_scratch_buf.get_buf_dev<T>(buffer_size);
-  T *host_dataset_buf = lazy_scratch_buf.get_buf_host<T>(buffer_size);
-
-  // write out regular cell-centered fields
-  for (const io::DatasetSpecEntry &cur_spec : dataset_spec.cc_dataset_entries) {
-    if constexpr (ForceF32Output) {
-      // todo: consider more robust behavior here
-      CHOLLA_ASSERT(cur_spec.condition == io::WriteCond::ALWAYS, "unexpected case");
-      CHOLLA_ASSERT(cur_spec.io_buf == field::IOBuf::DEVICE, "unexpected case");
-      Real *ptr = &G.C.device[cur_spec.field_id * H.n_cells];
-      Write_HDF5_Field_3D(H.nx, H.ny, nx_dset, ny_dset, nz_dset, H.n_ghost, file_id, host_dataset_buf, dev_dataset_buf,
-                          ptr, cur_spec.name.c_str());
-    } else {
-      if (cur_spec.condition == io::WriteCond::REQUIRE_COMPLETE_DATA && not H.Output_Complete_Data) continue;
-      if (cur_spec.io_buf == field::IOBuf::HOST) {
-        Real *ptr = &G.C.host[cur_spec.field_id * H.n_cells];
-        Write_Grid_HDF5_Field_CPU(H, file_id, host_dataset_buf, ptr, cur_spec.name.c_str());
-      } else {
-        Real *ptr = &G.C.device[cur_spec.field_id * H.n_cells];
-        Write_Grid_HDF5_Field_GPU(H, file_id, host_dataset_buf, dev_dataset_buf, ptr, cur_spec.name.c_str());
-      }
-    }
-  }
-
-  // write out magnetic fields
-  // -> we maintain historical behavior and only do this for 3D simulations
-  if (is_3D) {
-    const char *dset_names[3] = {"/magnetic_x", "/magnetic_y", "/magnetic_z"};
-    for (int i = 0; i < 3; i++) {
-      if (not dataset_spec.write_mag[i]) continue;
-      const char *field_name = dset_names[i] + 1;
-      Real *ptr              = &G.C.device[H.n_cells * G.field_info.field_id(field_name).value()];
-      if constexpr (ForceF32Output) {
-        // TODO (by Alwin, for anyone) : Repair output format if needed and remove the chprintf when appropriate
-        chprintf("WARNING: MHD float-32 output has a different output format than float-64\n");
-        Write_HDF5_Field_3D(H.nx, H.ny, nx_dset + 1, ny_dset + 1, nz_dset + 1, H.n_ghost - 1, file_id, host_dataset_buf,
-                            dev_dataset_buf, ptr, dset_names[i]);
-      } else {
-        if (not H.Output_Complete_Data) continue;
-        int real_shape[3] = {H.nx_real + (i == 0), H.ny_real + (i == 1), H.nz_real + (i == 2)};
-        Write_HDF5_Field_3D(H.nx, H.ny, real_shape[0], real_shape[1], real_shape[2], H.n_ghost, file_id,
-                            host_dataset_buf, dev_dataset_buf, ptr, dset_names[i], i);
-      }
-    }
-  }
-
-  // handle all of the weird special cases
-  if constexpr (not ForceF32Output) {
-  #if defined(OUTPUT_TEMPERATURE) && defined(CHEMISTRY_GPU)
-    Compute_Gas_Temperature(G.Chem.Fields.temperature_h, false);
-    Write_Grid_HDF5_Field_CPU(H, file_id, host_dataset_buf, G.Chem.Fields.temperature_h, "/temperature");
-  #elif defined(OUTPUT_TEMPERATURE) && defined(COOLING_GRACKLE)
-    Write_Grid_HDF5_Field_CPU(H, file_id, host_dataset_buf, G.Cool.temperature, "/temperature");
-  #endif
-
-  #if defined(GRAVITY) && defined(OUTPUT_POTENTIAL)
-    if (is_3D) {  // 3D case
-      const Grav3D &Grav = G.Grav;
-      Write_Generic_HDF5_Field_GPU(Grav.nx_local + 2 * N_GHOST_POTENTIAL, Grav.ny_local + 2 * N_GHOST_POTENTIAL,
-                                   Grav.nz_local + 2 * N_GHOST_POTENTIAL, Grav.nx_local, Grav.ny_local, Grav.nz_local,
-                                   N_GHOST_POTENTIAL, file_id, host_dataset_buf, dev_dataset_buf, Grav.F.potential_d,
-                                   "/grav_potential");
-    }
-  #endif  // GRAVITY and OUTPUT_POTENTIAL
-  }
-
-  // close the file
-  herr_t status = H5Fclose(file_id);
-
-  if (status < 0) {
-    printf("File write failed.\n");
-    exit(-1);
-  }
-#endif
-}
-
-void io::FieldWriter::operator()(Grid3D &G, Parameters P, int nfile, const FnameTemplate &fname_template) const
-{
-  // create the filename
-  std::string filename = fname_template.format_fname(nfile, "");
-
-#if !defined(BINARY) && !defined(HDF5)
-  if (G.H.nx * G.H.ny * G.H.nz > 1000) printf("Ascii outputs only recommended for small problems!\n");
-#endif
-
-// open the file for binary writes
-#if defined BINARY
-  FILE *out;
-  out = fopen(filename.data(), "w");
-  if (out == NULL) {
-    printf("Error opening output file.\n");
-    exit(-1);
-  }
-
-  // write the header to the output file
-  G.Write_Header_Binary(out);
-
-  // write the conserved variables to the output file
-  G.Write_Grid_Binary(out);
-
-  // close the output file
-  fclose(out);
-
-// create the file for hdf5 writes
-#elif defined HDF5
-  Write_Fields_to_HDF5_helper_<false>(filename, G, this->h5_dataset_spec_, *this->lazy_scratch_buf_);
-
-#else
-  // open the file for txt writes
-  FILE *out;
-  out = fopen(filename.data(), "w");
-  if (out == NULL) {
-    printf("Error opening output file.\n");
-    exit(-1);
-  }
-
-  // write the header to the output file
-  G.Write_Header_Text(out);
-
-  // write the conserved variables to the output file
-  G.Write_Grid_Text(out);
-
-  // close the output file
-  fclose(out);
-#endif
-}
-
-void io::F32FieldWriter::operator()(Grid3D &G, Parameters P, int nfile, const FnameTemplate &fname_template) const
-{
-#ifdef HDF5
-  std::string filename = fname_template.format_fname(nfile, ".float32");
-  Write_Fields_to_HDF5_helper_<true>(filename, G, this->dataset_spec_, *this->lazy_scratch_buf_);
-#endif  // HDF5
 }
 
 /* Output a projection of the grid data to file. */
